Move visualize graph out of main
