+++ tierkreis/tierkreis/cli/tkr.py
@@ -182,11 +182,6 @@ def main() -> None:
         logging.warning(
             "To install it, please run 'pip install tierkreis-visualization'"
         )
-    except RuntimeError:
-        logging.warning("Runtime Error during tierkreis visualization.")
-        logging.warning(
-            "Please Make sure your installation is valid, and assets are provided.'"
-        )
     args = parser.parse_args(args=None if sys.argv[1:] else ["--help"])
     args.func(args)
