+++ tierkreis_visualization/tierkreis_visualization/app.py
@@ -1,11 +1,12 @@
 from sys import argv
-from uuid import UUID
 from tierkreis.controller.data.graph import GraphData
-from tierkreis.controller.storage.graphdata import GraphDataStorage
-from tierkreis.controller.storage.protocol import ControllerStorage
 from tierkreis_visualization.app_config import App, StorageType, lifespan
 from tierkreis_visualization.config import CONFIG
-from tierkreis_visualization.storage import file_storage_fn, graph_data_storage_fn
+from tierkreis_visualization.storage import (
+    file_storage_fn,
+    from_graph_data_storage_fn,
+    graph_data_storage_fn,
+)
 from tierkreis_visualization.routers.frontend import assets
 from tierkreis_visualization.routers.workflows import router as workflows_router
 from tierkreis_visualization.routers.frontend import router as frontend_router
@@ -41,10 +42,6 @@ def get_graph_data_app():
 
 def app_from_graph_data(graph_data: GraphData):
     app = get_app()
-
-    def inner(workflow_id: UUID) -> ControllerStorage:
-        return GraphDataStorage(UUID(int=0), graph=graph_data)
-
-    app.state.get_storage_fn = inner
+    app.state.get_storage_fn = from_graph_data_storage_fn(graph_data)
     app.state.storage_type = StorageType.GRAPHDATA
     return app
