+++ tierkreis_visualization/tierkreis_visualization/main.py
@@ -1,10 +1,6 @@
 from sys import argv
 
 import uvicorn
-from tierkreis.builder import GraphBuilder
-from tierkreis.controller.data.graph import GraphData
-
-from tierkreis_visualization.app import app_from_graph_data
 
 
 def start() -> None:
@@ -26,18 +22,6 @@ def graph() -> None:
         reload=True,
         reload_includes=reload_path,
     )
-
-
-def visualize_graph(graph_data: GraphData | GraphBuilder) -> None:
-    """Visualize a computation graph in a web browser.
-
-    :param graph_data: The computation graph to visualize.
-    :type graph_data: GraphData | GraphBuilder
-    """
-    if isinstance(graph_data, GraphBuilder):
-        graph_data = graph_data.get_data()
-    app = app_from_graph_data(graph_data)
-    uvicorn.run(app)
 
 
 if __name__ == "__main__":
